Quest updates vol 1, mostly text (Fighters Guild)
This is an update based on player feedback. Most of them are changes to dialogue messages only. The gameplay changes are:

M0B1XY01: The spawn rate for these enemies was set very high in 2017, and player feedback says they're overwhelming. Reduces to about 1/3 of their previous speed, which is plenty to stay engaged.

M0B20Y02: Same as above.

M0B00Y16: Changed spawning rate for the six helper giants so they appear every 75 seconds instead of every 6 seconds, due to this being a low-level quest.