Increases rep requirement for The Lost Child quest
The Lost Child K0C00Y05 is an innkeeper quest that sends you to a dungeon that pretends to have an objective. Players go there, waste hours not finding it, and finally find out that it wasn't the intended solution. This change won't remove the quest, but it'll ensure that people won't run into it as their first quest. Now they have to have 30 or higher reputation with the faction in order to receive this quest.